requirements.txt
python-telegram-bot==13.15
Flask==2.2.5
apscheduler==3.10.4
pytz==2024.1